+++ source/dotnet/Services/WebApi.IntegrationTests/WebApi/V2/SettlementReportControllerTests.cs
@@ -12,30 +12,36 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Net;
+using System.Text;
+using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
+using Energinet.DataHub.Wholesale.Application.SettlementReport;
+using Energinet.DataHub.Wholesale.Application.SettlementReport.Model;
 using Energinet.DataHub.Wholesale.WebApi.IntegrationTests.Fixtures.TestCommon.Fixture.WebApi;
 using Energinet.DataHub.Wholesale.WebApi.IntegrationTests.Fixtures.WebApi;
 using FluentAssertions;
+using Moq;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Energinet.DataHub.Wholesale.WebApi.IntegrationTests.WebApi;
+namespace Energinet.DataHub.Wholesale.WebApi.IntegrationTests.WebApi.V2;
 
 [Collection(nameof(WholesaleWebApiCollectionFixture))]
-public class SwaggerTests :
+public class SettlementReportControllerTests :
     WebApiTestBase<WholesaleWebApiFixture>,
     IClassFixture<WholesaleWebApiFixture>,
     IClassFixture<WebApiFactory>,
     IAsyncLifetime
 {
     private readonly HttpClient _client;
+    private readonly WebApiFactory _factory;
 
-    public SwaggerTests(
+    public SettlementReportControllerTests(
         WholesaleWebApiFixture wholesaleWebApiFixture,
         WebApiFactory factory,
         ITestOutputHelper testOutputHelper)
         : base(wholesaleWebApiFixture, testOutputHelper)
     {
+        _factory = factory;
         _client = factory.CreateClient();
     }
 
@@ -50,14 +56,25 @@ public class SwaggerTests :
         return Task.CompletedTask;
     }
 
-    [Fact]
-    public async Task Given_WebAPI_When_GettingSwaggerUi_Then_ReturnsHtmlPage()
+    [Theory]
+    [InlineAutoMoqData]
+    public async Task GetAsync_ReturnsSettlementReport(
+        Mock<ISettlementReportApplicationService> mock,
+        string settlementReportContent)
     {
+        // Arrange
+        var batchId = Guid.NewGuid();
+        using var stream = new MemoryStream(Encoding.UTF8.GetBytes(settlementReportContent));
+        var settlementReport = new SettlementReportDto(stream);
+        mock
+            .Setup(service => service.GetSettlementReportAsync(batchId))
+            .ReturnsAsync(settlementReport);
+        _factory.SettlementReportApplicationServiceMock = mock;
+
         // Act
-        var actualResponse = await _client.GetAsync("/swagger/index.html");
+        var actualContent = await _client.GetStringAsync($"/v2.3/settlementreport?batchId={batchId.ToString()}");
 
         // Assert
-        actualResponse.Content.Headers.ContentType!.ToString().Should().StartWith("text/html");
-        actualResponse.StatusCode.Should().Be(HttpStatusCode.OK);
+        actualContent.Should().Be(settlementReportContent);
     }
 }
